refactor: Move randomBetween() function to `cypress/support/index.ts` file and update its respective imports
