+++ cypress/support/commands.ts
@@ -58,14 +58,3 @@ Cypress.Commands.add(
     }
   }
 );
-
-/**
- * @description Returns a random number given two boundaries
- * @param min {Number} - lower boundary
- * @param max {Number} - upper bondary
- * @returns {Number}
- */
-export const randomBetween = (min: number, max: number) =>
-  Math.floor(Math.random() * max) + min;
-
-export const SITE_URL = `${Cypress.env("siteUrl")}`;
